* 'dev' of https://github.com/miniProject-cloneCoding/tinder:
  #5 조회 기능 구현 - JPQL로 변경

+++ src/main/java/clonecoding/tinder/member/repository/CustomMemberRepository.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface CustomMemberRepository {
 
-    List<Member> findAllWithoutLike(Long myId, Long offset, int limit);
+    List<Member> findAllWithoutLike(Long myId, int offset, int limit);
 }
